api/requirements.txt
certifi==2021.10.8
click==8.0.4
Flask==2.0.3
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
minio==7.1.5
psycopg2-binary==2.9.3
pycryptodome==3.14.1
PyJWT==2.3.0
SQLAlchemy==1.4.32
urllib3==1.26.8
Werkzeug==2.0.3